Add files via upload
Objective:
To estimate the compressive strength of concrete based on components and age using multilinear regression.

Data:
Obtained from UCI Machine Learning Repository: Concrete Compressive Strength Data Set and used in I-Cheng Yeh, "Modeling of strength of high performance concrete using artificial neural networks," Cement and Concrete Research, Vol. 28, No. 12, pp. 1797-1808 (1998). 

Problem Statement:
Structural engineers need to know the strength of materials in order to design safe structures. As a machine learning developer working for an engineering company, your job is to analyze data to develop a model that predicts the compressive strength of concrete based on what it is made of and how old the concrete is. The model must also provide an uncertainty in the strength prediction so that engineers can ensure that the design is safe.